+++ src/config/config.test.ts
@@ -1,5 +1,15 @@
 import config from './config';
 
+jest.mock('webextension-polyfill', () => {
+  return {
+    runtime: {
+      getManifest: () => {
+        return { version: 'test-version' };
+      },
+    },
+  };
+});
+
 describe('wakatime config', () => {
   it('snapshot of config', () => {
     expect(config).toMatchInlineSnapshot(`
@@ -23,12 +33,32 @@ describe('wakatime config', () => {
         },
         "currentUserApiUrl": "https://wakatime.com/api/v1/users/current",
         "detectionIntervalInSeconds": 60,
+        "devSites": "https://codepen.io/
+      https://www.codewars.com/
+      https://dev.to/
+      https://github.com/
+      https://www.hackerrank.com/
+      https://leetcode.com/
+      https://developer.mozilla.org/en-US/
+      https://stackoverflow.com/
+      https://www.udemy.com/
+      https://www.w3schools.com/",
         "heartbeatApiUrl": "https://wakatime.com/api/v1/users/current/heartbeats",
         "loggingEnabled": true,
         "loggingStyle": "blacklist",
         "loggingType": "domain",
         "logoutUserUrl": "https://wakatime.com/logout",
         "name": "WakaTime",
+        "socialMediaSites": "https://www.facebook.com/
+      https://www.instagram.com/
+      https://www.linkedin.com/
+      https://www.pinterest.com/
+      https://www.reddit.com/
+      https://www.snapchat.com/
+      https://www.tiktok.com/
+      https://twitter.com/
+      https://www.whatsapp.com/
+      https://www.youtube.com/",
         "states": [
           "allGood",
           "notLogging",
@@ -45,6 +75,7 @@ describe('wakatime config', () => {
           "notSignedIn": "Not signed In",
           "whitelisted": "This URL is not on your whitelist",
         },
+        "trackSocialMedia": true,
         "version": "test-version",
       }
     `);