+++ src/config/config.ts
@@ -1,3 +1,5 @@
+import browser from 'webextension-polyfill';
+
 /**
  * Logging
  */
@@ -62,6 +64,8 @@ export interface Config {
    * no activity in the browser for x second
    */
   detectionIntervalInSeconds: number;
+
+  devSites: string;
   /**
    * Url to which to send the heartbeat
    */
@@ -80,6 +84,7 @@ export interface Config {
    * Extension name
    */
   name: string;
+  socialMediaSites: string;
   states: ApiStates[];
   /**
    * Get stats from the wakatime api
@@ -90,6 +95,7 @@ export interface Config {
    */
   theme: 'light';
   tooltips: Tooltips;
+  trackSocialMedia: boolean;
   /**
    * Version of the extension
    */
@@ -122,6 +128,9 @@ const config: Config = {
 
   detectionIntervalInSeconds: 60,
 
+  devSites:
+    'https://codepen.io/\nhttps://www.codewars.com/\nhttps://dev.to/\nhttps://github.com/\nhttps://www.hackerrank.com/\nhttps://leetcode.com/\nhttps://developer.mozilla.org/en-US/\nhttps://stackoverflow.com/\nhttps://www.udemy.com/\nhttps://www.w3schools.com/',
+
   heartbeatApiUrl:
     process.env.HEART_BEAT_API_URL ?? 'https://wakatime.com/api/v1/users/current/heartbeats',
 
@@ -134,6 +143,8 @@ const config: Config = {
   logoutUserUrl: process.env.LOGOUT_USER_URL ?? 'https://wakatime.com/logout',
 
   name: 'WakaTime',
+
+  socialMediaSites: `https://www.facebook.com/\nhttps://www.instagram.com/\nhttps://www.linkedin.com/\nhttps://www.pinterest.com/\nhttps://www.reddit.com/\nhttps://www.snapchat.com/\nhttps://www.tiktok.com/\nhttps://twitter.com/\nhttps://www.whatsapp.com/\nhttps://www.youtube.com/`,
 
   states: ['allGood', 'notLogging', 'notSignedIn', 'blacklisted', 'whitelisted'],
 
@@ -149,6 +160,7 @@ const config: Config = {
     notSignedIn: 'Not signed In',
     whitelisted: 'This URL is not on your whitelist',
   },
+  trackSocialMedia: true,
 
   version: browser.runtime.getManifest().version,
 };